requirements.txt
playwright==1.48.0
pytest-playwright==0.5.2
pytest==8.3.3
python-dotenv==1.0.1